Move test types
