+++ openapi/generator_test.go
@@ -3,8 +3,8 @@ package openapi
 import (
 	"encoding/json"
 	"fmt"
-	"github.com/wI2L/fizz/testdata/test_types"
-	"github.com/wI2L/fizz/testdata/test_types_extra"
+	baseTypes "github.com/wI2L/fizz/openapi/test_types/base_types"
+	extraTypes "github.com/wI2L/fizz/openapi/test_types/extra_types"
 	"io/ioutil"
 	"math"
 	"reflect"
@@ -142,7 +142,7 @@ func TestSchemaFromMapWithUnsupportedKeys(t *testing.T) {
 func TestSchemaFromDuplicatedType(t *testing.T) {
 	g := gen(t)
 
-	schema := g.newSchemaFromType(rt(new(test_types_extra.D)))
+	schema := g.newSchemaFromType(rt(new(extraTypes.D)))
 	assert.NotNil(t, schema)
 	assert.Len(t, g.Errors(), 1)
 	assert.Implements(t, (*error)(nil), g.Errors()[0])
@@ -155,7 +155,7 @@ func TestSchemaFromComplex(t *testing.T) {
 	g := gen(t)
 	g.UseFullSchemaNames(false)
 
-	sor := g.newSchemaFromType(rt(new(test_types.X)))
+	sor := g.newSchemaFromType(rt(new(baseTypes.X)))
 	assert.NotNil(t, sor)
 
 	b, err := json.Marshal(sor)
@@ -430,7 +430,7 @@ func TestAddOperation(t *testing.T) {
 			},
 		},
 	}
-	_, err := g.AddOperation(path, "POST", "Test", reflect.TypeOf(&In{}), reflect.TypeOf(test_types.Z{}), infos)
+	_, err := g.AddOperation(path, "POST", "Test", reflect.TypeOf(&In{}), reflect.TypeOf(baseTypes.Z{}), infos)
 	if err != nil {
 		t.Error(err)
 	}
@@ -473,7 +473,7 @@ func TestAddOperation(t *testing.T) {
 	}
 	// Try to add the operation again with the same
 	// identifier. Expected to fail.
-	_, err = g.AddOperation(path, "POST", "Test", reflect.TypeOf(&In{}), reflect.TypeOf(test_types.Z{}), infos)
+	_, err = g.AddOperation(path, "POST", "Test", reflect.TypeOf(&In{}), reflect.TypeOf(baseTypes.Z{}), infos)
 	assert.NotNil(t, err)
 
 	// Add an operation with a bad input type.
@@ -489,25 +489,25 @@ func TestTypeName(t *testing.T) {
 		t.Error(err)
 	}
 	// Typer interface.
-	name := g.typeName(rt(new(test_types.X)))
+	name := g.typeName(rt(new(baseTypes.X)))
 	assert.Equal(t, "XXX", name)
 
 	// Override. This has precedence
 	// over the interface implementation.
-	err = g.OverrideTypeName(rt(new(test_types.X)), "")
+	err = g.OverrideTypeName(rt(new(baseTypes.X)), "")
 	assert.NotNil(t, err)
-	assert.Equal(t, "XXX", g.typeName(rt(new(test_types.X))))
+	assert.Equal(t, "XXX", g.typeName(rt(new(baseTypes.X))))
 
-	g.OverrideTypeName(rt(new(test_types.X)), "xXx")
-	assert.Equal(t, "xXx", g.typeName(rt(test_types.X{})))
+	g.OverrideTypeName(rt(new(baseTypes.X)), "xXx")
+	assert.Equal(t, "xXx", g.typeName(rt(baseTypes.X{})))
 
-	err = g.OverrideTypeName(rt(new(test_types.X)), "YYY")
+	err = g.OverrideTypeName(rt(new(baseTypes.X)), "YYY")
 	assert.NotNil(t, err)
 
 	// Default.
-	assert.Equal(t, "Test_typesY", g.typeName(rt(new(test_types.Y))))
+	assert.Equal(t, "Base_typesY", g.typeName(rt(new(baseTypes.Y))))
 	g.UseFullSchemaNames(false)
-	assert.Equal(t, "Y", g.typeName(rt(test_types.Y{})))
+	assert.Equal(t, "Y", g.typeName(rt(baseTypes.Y{})))
 
 	// Unnamed type.
 	assert.Equal(t, "", g.typeName(rt(struct{}{})))
@@ -673,18 +673,18 @@ func TestOverrideSchema(t *testing.T) {
 	g := gen(t)
 
 	// Type is mandatory.
-	err := g.OverrideDataType(rt(test_types.W{}), "", "wallet")
+	err := g.OverrideDataType(rt(baseTypes.W{}), "", "wallet")
 	assert.NotNil(t, err)
 
 	// Success.
-	err = g.OverrideDataType(rt(&test_types.W{}), "string", "wallet")
+	err = g.OverrideDataType(rt(&baseTypes.W{}), "string", "wallet")
 	assert.Nil(t, err)
 
 	// Data type already overidden.
-	err = g.OverrideDataType(rt(&test_types.W{}), "string", "wallet")
+	err = g.OverrideDataType(rt(&baseTypes.W{}), "string", "wallet")
 	assert.NotNil(t, err)
 
-	sor := g.newSchemaFromType(rt(test_types.W{}))
+	sor := g.newSchemaFromType(rt(baseTypes.W{}))
 	assert.NotNil(t, sor)
 
 	schema := g.resolveSchema(sor)
