+++ openapi/test_types/base_types/types.go
@@ -1,4 +1,4 @@
-package test_types
+package base_types
 
 import "time"
