--- testdata/test_types_extra/types.go
@@ -1,12 +1,0 @@
-package test_types_extra
-
-import "github.com/wI2L/fizz/testdata/test_types"
-
-type W struct {
-	C, D int
-}
-
-type D struct {
-	W  W
-	Wd test_types.W
-}
